EntlibLogger tests. Refactor.
