+++ src/LoggingIntegrationTests/App.config
@@ -1,0 +1,49 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <configSections>
+    <section name="loggingConfiguration" type="Microsoft.Practices.EnterpriseLibrary.Logging.Configuration.LoggingSettings, Microsoft.Practices.EnterpriseLibrary.Logging" />
+  </configSections>
+  <loggingConfiguration name="Logging Application Block" tracingEnabled="false" defaultCategory="General" logWarningsWhenNoCategoriesMatch="true">
+    <listeners>
+      <add source="Enterprise Library Logging" formatter="" log="Application" machineName="" listenerDataType="Microsoft.Practices.EnterpriseLibrary.Logging.Configuration.FormattedEventLogTraceListenerData, Microsoft.Practices.EnterpriseLibrary.Logging" traceOutputOptions="None" filter="All" type="Microsoft.Practices.EnterpriseLibrary.Logging.TraceListeners.FormattedEventLogTraceListener, Microsoft.Practices.EnterpriseLibrary.Logging" name="Application EventLog TraceListener" />
+      <add databaseInstanceName="SnCrMsSql" writeLogStoredProcName="proc_LogWrite" addCategoryStoredProcName="proc_LogAddCategory" formatter="SnXmlFormatter" listenerDataType="Microsoft.Practices.EnterpriseLibrary.Logging.Database.Configuration.FormattedDatabaseTraceListenerData, Microsoft.Practices.EnterpriseLibrary.Logging.Database" traceOutputOptions="None" filter="All" type="Microsoft.Practices.EnterpriseLibrary.Logging.Database.FormattedDatabaseTraceListener, Microsoft.Practices.EnterpriseLibrary.Logging.Database" name="Database Trace Listener" />
+      <add listenerDataType="Microsoft.Practices.EnterpriseLibrary.Logging.Configuration.CustomTraceListenerData, Microsoft.Practices.EnterpriseLibrary.Logging" traceOutputOptions="None" filter="All"
+           type="SenseNet.Diagnostics.OneLineTraceListener, SenseNet.EntlibLogger" name="Debug Trace Listener" initializeData="" formatter="OneLineFormatter" />
+      <add source="SenseNetInstrumentation" formatter="Text Formatter" log="SenseNet" machineName="" listenerDataType="Microsoft.Practices.EnterpriseLibrary.Logging.Configuration.FormattedEventLogTraceListenerData, Microsoft.Practices.EnterpriseLibrary.Logging" traceOutputOptions="None" filter="Information" type="Microsoft.Practices.EnterpriseLibrary.Logging.TraceListeners.FormattedEventLogTraceListener, Microsoft.Practices.EnterpriseLibrary.Logging" name="SenseNet EventLog TraceListener" />
+    </listeners>
+    <formatters>
+      <add template="{severity} {category} {message}: Properties: {dictionary({key}={value};)}; Timestamp: {timestamp}; Machine: {machine}; AppDomain: {appDomain}; ProcessId: {processId}; Process: {processName}; Win32ThreadId: {win32ThreadId}; ThreadName: {threadName}" type="Microsoft.Practices.EnterpriseLibrary.Logging.Formatters.TextFormatter, Microsoft.Practices.EnterpriseLibrary.Logging" name="OneLineFormatter" />
+      <add template="Timestamp: {timestamp}&#xD;&#xA;Message: {message}&#xD;&#xA;Category: {category}&#xD;&#xA;Priority: {priority}&#xD;&#xA;EventId: {eventid}&#xD;&#xA;Severity: {severity}&#xD;&#xA;Title:{title}&#xD;&#xA;Machine: {machine}&#xD;&#xA;Application Domain: {appDomain}&#xD;&#xA;Process Id: {processId}&#xD;&#xA;Process Name: {processName}&#xD;&#xA;Win32 Thread Id: {win32ThreadId}&#xD;&#xA;Thread Name: {threadName}&#xD;&#xA;Extended Properties: {dictionary({key} - {value}&#xD;&#xA;)}" type="Microsoft.Practices.EnterpriseLibrary.Logging.Formatters.TextFormatter, Microsoft.Practices.EnterpriseLibrary.Logging" name="Text Formatter" />
+      <add type="SenseNet.Diagnostics.SnXmlFormatter, SenseNet.EntlibLogger" name="SnXmlFormatter" />
+    </formatters>
+    <categorySources>
+      <add switchValue="All" name="Audit">
+        <listeners>
+          <add name="Database Trace Listener" />
+          <add name="Debug Trace Listener" />
+        </listeners>
+      </add>
+      <add switchValue="All" name="General">
+        <listeners>
+          <add name="Debug Trace Listener" />
+        </listeners>
+      </add>
+    </categorySources>
+    <specialSources>
+      <allEvents switchValue="All" name="All Events">
+        <listeners>
+          <add name="SenseNet EventLog TraceListener" />
+        </listeners>
+      </allEvents>
+      <notProcessed switchValue="All" name="Unprocessed Category" />
+      <errors switchValue="All" name="Logging Errors &amp; Warnings">
+        <listeners>
+          <add name="SenseNet EventLog TraceListener" />
+        </listeners>
+      </errors>
+    </specialSources>
+  </loggingConfiguration>
+  <connectionStrings>
+    <add name="SnCrMsSql" connectionString="Initial Catalog=sn7tests;Data Source=.\SQL2016;Integrated Security=SSPI;Persist Security Info=False" providerName="System.Data.SqlClient" />
+  </connectionStrings>
+</configuration>
