+++ src/LoggingIntegrationTests/LoggingToEventLogTests.cs
@@ -1,20 +1,21 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
-using System.Text;
+using LoggingIntegrationTests.Implementations;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SenseNet.Diagnostics;
 
 namespace LoggingIntegrationTests
 {
     [TestClass]
-    public class LoggingToEventLogTests
+    public class LoggingToEventLogTests : LoggerTestBase
     {
+        private static readonly string LogName = "SenseNet";
+        private static readonly string LogSource = "SenseNetInstrumentation";
+
         [TestMethod]
         public void Logging_Audit_ToEventLog()
         {
-            using (new SnEventLoggerSwindler())
+            using (SwindleLogger(new SnEventLogger(LogName, LogSource)))
             {
                 var testValue = Guid.NewGuid().ToString();
 
@@ -22,7 +23,7 @@ namespace LoggingIntegrationTests
                 SnLog.WriteAudit(new TestAuditEvent(testValue));
 
                 // assert
-                var lastEntry = GetLastEntry();
+                var lastEntry = GetLastEventLogEntry();
                 var entryData = ParseEventlogEntryData(lastEntry.Message);
 
                 Assert.AreEqual(testValue, entryData["Message"]);
@@ -35,7 +36,7 @@ namespace LoggingIntegrationTests
         [TestMethod]
         public void Logging_Information_ToEventLog()
         {
-            using (new SnEventLoggerSwindler())
+            using (SwindleLogger(new SnEventLogger(LogName, LogSource)))
             {
                 var testMessage = Guid.NewGuid().ToString();
 
@@ -43,7 +44,7 @@ namespace LoggingIntegrationTests
                 SnLog.WriteInformation(testMessage);
 
                 // assert
-                var lastEntry = GetLastEntry();
+                var lastEntry = GetLastEventLogEntry();
                 var entryData = ParseEventlogEntryData(lastEntry.Message);
 
                 Assert.AreEqual(testMessage, entryData["Message"]);
@@ -51,58 +52,6 @@ namespace LoggingIntegrationTests
                 Assert.AreEqual("Information", entryData["Severity"]);
                 Assert.AreEqual(EventLogEntryType.Information, lastEntry.EntryType);
             }
-        }
-
-        /* ============================================================= */
-
-        private class SnEventLoggerSwindler : IDisposable
-        {
-            private readonly IEventLogger _backup;
-            public SnEventLoggerSwindler()
-            {
-                _backup = SnLog.Instance;
-                SnLog.Instance = new SnEventLogger("SenseNet", "SenseNetInstrumentation");
-            }
-            public void Dispose()
-            {
-                SnLog.Instance = _backup;
-            }
-        }
-
-        private EventLogEntry GetLastEntry()
-        {
-            var logs = EventLog.GetEventLogs();
-            var log = logs.FirstOrDefault(l => l.LogDisplayName == "SenseNet");
-            Assert.IsNotNull(log);
-            var entries = new List<EventLogEntry>();
-            foreach (EventLogEntry entry in log.Entries)
-                entries.Add(entry);
-            return entries.Last();
-        }
-
-        private Dictionary<string, string> ParseEventlogEntryData(string text)
-        {
-            var result = new Dictionary<string, string>();
-            var fields = text.Split(new[] { Environment.NewLine }, StringSplitOptions.RemoveEmptyEntries);
-            var index = 0;
-            while (true)
-            {
-                var field = fields[index++];
-                var p = field.IndexOf(':');
-                var name = field.Substring(0, p);
-                var value = field.Length > p ? field.Substring(p + 1).Trim() : string.Empty;
-                if (name != "Extended Properties")
-                {
-                    result.Add(name, value);
-                    continue;
-                }
-                var extendedValue = new StringBuilder(value);
-                for (int i = index; i < fields.Length; i++)
-                    extendedValue.Append(", ").Append(fields[i]);
-                result.Add(name, extendedValue.ToString());
-                break;
-            }
-            return result;
         }
     }
 }
